Rename files to be camel case
