+++ src/extensions/loggingExtension.ts
